services/encoder_service/requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
numpy
redis